[UPDATE] Change to PosgreSQL, fix entities

+++ src/models/product_model.js
@@ -19,7 +19,7 @@ const Product = sequelize.define(
       allowNull: false,
     },
     price: {
-      type: DataTypes.FLOAT,
+      type: DataTypes.DECIMAL(12, 2),
       allowNull: false,
     },
     stock: {
